simplify readability with BaseRequestOptions
