+++ src/RESTController.ts
@@ -5,6 +5,15 @@ import ParseError from './ParseError';
 import { resolvingPromise } from './promiseUtils';
 import { polyfillFetch } from './Xhr.weapp';
 
+/** Base options for all Parse requests that go through _getRequestOptions */
+export interface BaseRequestOptions {
+  useMasterKey?: boolean;
+  useMaintenanceKey?: boolean;
+  sessionToken?: string;
+  installationId?: string;
+  context?: Record<string, unknown>;
+}
+
 export interface RequestOptions {
   useMasterKey?: boolean;
   useMaintenanceKey?: boolean;
