+++ types/ParseObject.d.ts
@@ -4,7 +4,7 @@ import ParseFile from './ParseFile';
 import { Op } from './ParseOp';
 import ParseRelation from './ParseRelation';
 import type { AttributeMap, OpsMap } from './ObjectStateMutations';
-import type { RequestOptions, FullOptions } from './RESTController';
+import type { RequestOptions, FullOptions, BaseRequestOptions } from './RESTController';
 import type ParseGeoPoint from './ParseGeoPoint';
 import type ParsePolygon from './ParsePolygon';
 export interface Pointer {
@@ -17,14 +17,6 @@ interface SaveParams {
     method: string;
     path: string;
     body: AttributeMap;
-}
-/** Base options for all Parse requests that go through _getRequestOptions */
-export interface BaseRequestOptions {
-    useMasterKey?: boolean;
-    useMaintenanceKey?: boolean;
-    sessionToken?: string;
-    installationId?: string;
-    context?: AttributeMap;
 }
 export interface SaveOptions extends BaseRequestOptions {
     /** If `false`, nested objects will not be saved (default is `true`). */
